order attachments in the field
separate views and models into their own files
